refactor: apply mine-challenge findings on handler-telemetry-debt
- job-detail.tsx: replace findIndex/splice string-matching for the Skipped
  stat cell with a declarative insertAfterCancelledOrTimedOut slot on
  buildCommonStatCells, eliminating the unenforced cross-file label contract;
  add position-assertion tests
- move SYNTHETIC_TIMESTAMP from web_telemetry_helpers.py to test_utils/config.py
  so the four web_*_helpers.py siblings no longer cross-import each other for
  a value already available from their shared config module
- move TestListenerSummaryLastErrorRowCoherence and
  TestJobSummaryLastErrorRowCoherence out of test_execution_queries.py into
  test_listener_queries.py / test_job_queries.py, next to the other
  get_listener_summary/get_job_summary tests they actually cover
- consolidate the duplicated 'Separation rationale' docstring across the 5
  schema domain files into a single copy in schemas/__init__.py
