+++ tests/integration/telemetry/test_job_queries.py
@@ -229,3 +229,172 @@ class TestGetJobSummarySinceScoped:
         assert row.total_executions == 2
         assert row.successful == 1
         assert row.failed == 1
+
+
+class TestJobSummaryLastErrorRowCoherence:
+    """Verify that last_error_* fields all come from the same job_executions row."""
+
+    async def test_multiple_errors_returns_most_recent(
+        self,
+        query_service: TelemetryQueryService,
+        db: tuple[DatabaseService, int],
+    ) -> None:
+        """Multiple errors at different timestamps — all error columns from the most recent row."""
+        db_svc, session_id = db
+        job_id = await insert_job(db_svc, job_name="multi_err_job")
+
+        base_ts = BASE_TS
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="error",
+            error_type="OldError",
+            error_message="old message",
+            error_traceback="old traceback",
+            execution_start_ts=base_ts + 1.0,
+        )
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="error",
+            error_type="NewError",
+            error_message="new message",
+            error_traceback="new traceback",
+            execution_start_ts=base_ts + 10.0,
+        )
+        # Success after errors — should not affect error fields
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="success",
+            execution_start_ts=base_ts + 20.0,
+        )
+
+        rows = await query_service.get_job_summary("test_app", 0)
+        assert len(rows) == 1
+        row = rows[0]
+        assert row.last_error_type == "NewError"
+        assert row.last_error_message == "new message"
+        assert row.last_error_traceback == "new traceback"
+        assert row.last_error_ts == pytest.approx(base_ts + 10.0)
+
+    async def test_single_error_returned(
+        self,
+        query_service: TelemetryQueryService,
+        db: tuple[DatabaseService, int],
+    ) -> None:
+        """Single error execution — all error columns are populated from that row."""
+        db_svc, session_id = db
+        job_id = await insert_job(db_svc, job_name="single_err_job")
+
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="error",
+            error_type="RuntimeError",
+            error_message="runtime boom",
+            error_traceback="tb: boom at line 1",
+            execution_start_ts=BASE_TS + 5.0,
+        )
+
+        rows = await query_service.get_job_summary("test_app", 0)
+        assert len(rows) == 1
+        row = rows[0]
+        assert row.last_error_type == "RuntimeError"
+        assert row.last_error_message == "runtime boom"
+        assert row.last_error_traceback == "tb: boom at line 1"
+
+    async def test_no_errors_returns_none(
+        self,
+        query_service: TelemetryQueryService,
+        db: tuple[DatabaseService, int],
+    ) -> None:
+        """No errors — all last_error_* fields are None."""
+        db_svc, session_id = db
+        job_id = await insert_job(db_svc, job_name="clean_job")
+
+        await insert_execution(db_svc, job_id, session_id, status="success")
+        await insert_execution(db_svc, job_id, session_id, status="success")
+
+        rows = await query_service.get_job_summary("test_app", 0)
+        assert len(rows) == 1
+        row = rows[0]
+        assert row.last_error_type is None
+        assert row.last_error_message is None
+        assert row.last_error_traceback is None
+        assert row.last_error_ts is None
+
+    async def test_since_filter_scopes_error_cte(
+        self,
+        query_service: TelemetryQueryService,
+        db: tuple[DatabaseService, int],
+    ) -> None:
+        """Error before the since window is excluded; error inside the window is returned."""
+        db_svc, session_id = db
+        job_id = await insert_job(db_svc, job_name="windowed_job")
+
+        base_ts = BASE_TS
+        since_ts = base_ts + 50.0
+
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="error",
+            error_type="OldError",
+            error_message="before window",
+            error_traceback="old tb",
+            execution_start_ts=base_ts + 1.0,
+        )
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="error",
+            error_type="NewError",
+            error_message="inside window",
+            error_traceback="new tb",
+            execution_start_ts=base_ts + 100.0,
+        )
+
+        rows = await query_service.get_job_summary("test_app", 0, since=since_ts)
+        assert len(rows) == 1
+        row = rows[0]
+        assert row.last_error_type == "NewError"
+        assert row.last_error_message == "inside window"
+        assert row.last_error_traceback == "new tb"
+
+    async def test_since_filter_excludes_all_errors_returns_none(
+        self,
+        query_service: TelemetryQueryService,
+        db: tuple[DatabaseService, int],
+    ) -> None:
+        """All errors before since window — last_error_* fields are None."""
+        db_svc, session_id = db
+        job_id = await insert_job(db_svc, job_name="stale_job")
+
+        base_ts = BASE_TS
+        since_ts = base_ts + 500.0
+
+        await insert_execution(
+            db_svc,
+            job_id,
+            session_id,
+            status="error",
+            error_type="StaleError",
+            error_message="stale",
+            error_traceback="stale tb",
+            execution_start_ts=base_ts + 1.0,
+        )
+
+        rows = await query_service.get_job_summary("test_app", 0, since=since_ts)
+        assert len(rows) == 1
+        row = rows[0]
+        assert row.last_error_type is None
+        assert row.last_error_message is None
+        assert row.last_error_traceback is None
+        assert row.last_error_ts is None
